+++ GRDB/Core/TransactionObserver.swift
@@ -737,8 +737,7 @@ public protocol TransactionObserver: AnyObject {
     ///
     /// **Availability Info**
     ///
-    /// Requires SQLite 3.13.0 +
-    /// Compiled with option SQLITE_ENABLE_PREUPDATE_HOOK
+    /// Requires SQLite compiled with option SQLITE_ENABLE_PREUPDATE_HOOK.
     ///
     /// As of OSX 10.11.5, and iOS 9.3.2, the built-in SQLite library
     /// does not have this enabled, so you'll need to compile your own